Move the GitHub API into it's own object
