+++ source/platforms/_tests/fixtures/static_file.md
@@ -1,0 +1,1 @@
+The All-Defector is a purported glitch in the Dilemma Prison that appears to prisoners as themselves. This gogol always defects, hence the name.
