+++ source/runner/Executor.ts
@@ -3,7 +3,7 @@ import { DangerDSL } from "../dsl/DangerDSL"
 import { CISource } from "../ci_source/ci_source"
 import { Platform } from "../platforms/platform"
 import { DangerResults } from "../dsl/DangerResults"
-import { template as githubResultsTemplate } from "./templates/github-issue-template"
+import { template as githubResultsTemplate } from "./templates/githubIssueTemplate"
 import { createDangerfileRuntimeEnvironment, runDangerfileEnvironment } from "./DangerfileRunner"
 import { DangerfileRuntimeEnv } from "./types"
 import * as debug from "debug"
